requirements.txt
# Core
google-cloud-bigquery
google-cloud-storage
db-dtypes
pandas<2.2
pyarrow

# PySpark
pyspark==3.5.3

# dbt
dbt-bigquery

# Airflow
apache-airflow==2.8.4
apache-airflow-providers-google

# Ingestion
requests
psycopg2-binary
